move CheckInAllocMsg to more logical place
